ETL-PROCESS/wait_until_ready/requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python
